* Wiring up EF calls as DataProvider * Added IOptions system with DI

+++ ConCode.NET.Database/Tables/SpeakerTalks.sql
@@ -1,6 +1,6 @@
 ﻿CREATE TABLE [dbo].[SpeakerTalks]
 (
 	[Id] BIGINT NOT NULL PRIMARY KEY IDENTITY (1, 1), 
-    [SpeakerId] INT NOT NULL, 
-    [TalkId] INT NOT NULL
+    [SpeakerId] BIGINT NOT NULL, 
+    [TalkId] BIGINT NOT NULL
 )
