+++ src/ConCode.NET.Core/Domain/SpeakerInfo.cs
@@ -8,11 +8,11 @@ namespace ConCode.NET.Core.Domain
 {
     public class SpeakerInfo
    {
-        public int Id { get; set; }
+        public long Id { get; set; }
         public IEnumerable<Talk> Talks { get; set; }
 
         public string Tagline { get; set; }
-        public int UserId { get; internal set; }
+        public long UserId { get; internal set; }
         public User User { get; internal set; }
     }
 }
